+++ genium/src/test/java/com/here/genium/generator/java/JavaValueMapperTest.kt
@@ -22,8 +22,11 @@ package com.here.genium.generator.java
 import com.here.genium.Genium
 import com.here.genium.generator.common.nameRuleSetFromConfig
 import com.here.genium.model.java.JavaCustomType
+import com.here.genium.model.java.JavaPrimitiveType
 import com.here.genium.model.java.JavaTemplateType
 import com.here.genium.model.java.JavaType
+import com.here.genium.model.lime.LimeBasicType
+import com.here.genium.model.lime.LimeBasicTypeRef
 import com.here.genium.model.lime.LimeContainer
 import com.here.genium.model.lime.LimeElement
 import com.here.genium.model.lime.LimeEnumeration
@@ -31,6 +34,7 @@ import com.here.genium.model.lime.LimeEnumerator
 import com.here.genium.model.lime.LimeEnumeratorRef
 import com.here.genium.model.lime.LimePath
 import com.here.genium.model.lime.LimeLazyTypeRef
+import com.here.genium.model.lime.LimeTypeDef
 import com.here.genium.model.lime.LimeValue
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
@@ -55,6 +59,28 @@ class JavaValueMapperTest {
         limeReferenceMap["bar"] = LimeEnumeration(LimePath(emptyList(), listOf("foo", "bar")))
         limeReferenceMap["baz"] = LimeEnumerator(LimePath(emptyList(), listOf("baz")))
         val enumeratorRef = LimeLazyTypeRef("bar", limeReferenceMap)
+        val valueRef = LimeEnumeratorRef(limeReferenceMap, "baz")
+        val limeValue = LimeValue.Enumerator(enumeratorRef, valueRef)
+
+        val result = valueMapper.mapValue(limeValue, javaType)
+
+        assertTrue(result.isCustom)
+        assertEquals("Foo.Bar.BAZ", result.name)
+    }
+
+    @Test
+    fun mapTypeDefToEnumerator() {
+        limeReferenceMap["foo"] = LimeContainer(
+            LimePath(emptyList(), listOf("foo")),
+            type = LimeContainer.ContainerType.INTERFACE
+        )
+        limeReferenceMap["bar"] = LimeEnumeration(LimePath(emptyList(), listOf("foo", "bar")))
+        limeReferenceMap["baz"] = LimeEnumerator(LimePath(emptyList(), listOf("baz")))
+        limeReferenceMap["barDef"] = LimeTypeDef(
+            LimePath(emptyList(), listOf("foo", "barDef")),
+            typeRef = LimeLazyTypeRef("bar", limeReferenceMap)
+        )
+        val enumeratorRef = LimeLazyTypeRef("barDef", limeReferenceMap)
         val valueRef = LimeEnumeratorRef(limeReferenceMap, "baz")
         val limeValue = LimeValue.Enumerator(enumeratorRef, valueRef)
 
@@ -125,5 +151,20 @@ class JavaValueMapperTest {
 
         assertTrue(result.isCustom)
         assertEquals("Foo", result.name)
+    }
+
+    @Test
+    fun mapTypedefToLongValue() {
+        limeReferenceMap["foo"] = LimeTypeDef(
+            LimePath(emptyList(), listOf("foo")),
+            typeRef = LimeBasicTypeRef(LimeBasicType.TypeId.INT64)
+        )
+        val limeTypeRef = LimeLazyTypeRef("foo", limeReferenceMap)
+        val limeValue = LimeValue.Literal(limeTypeRef, "1")
+        val javaType = JavaPrimitiveType.LONG
+
+        val result = valueMapper.mapValue(limeValue, javaType)
+
+        assertEquals("1L", result.name)
     }
 }